third_party/blink/web_tests/fast/events/check-defocus-event-order-when-triggered-by-mouse-click-expected.txt

This test checks that when a user focuses on the "first input" text field, modifies its contents, and then defocuses it by clicking on the "second input" text field that the following DOM events are fired in order: Focus, Change, Blur, Focus (on "second input"), Click (on "second input"). Note, this test must be run by Dump Render Tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
PASS successfullyParsed is true

TEST COMPLETE

PASS fired event is (firstInput, focus).
PASS fired event is (firstInput, change).
PASS fired event is (firstInput, blur).
PASS fired event is (secondInput, focus).
PASS fired event is (secondInput, click).

TEST COMPLETE
